requirements.txt
boto3==1.17.53
botocore==1.20.53
jmespath==0.10.0
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
s3transfer==0.3.7
six==1.15.0
urllib3==1.26.4
